License V1 and V2 combined to a universal license.
SystemId has now 4 features it can be bound to.

+++ common/src/main/java/de/shadowsoft/greenLicense/common/license/generator/Selector.java
@@ -3,15 +3,15 @@ package de.shadowsoft.greenLicense.common.license.generator;
 
 public class Selector {
 
-    private final int id;
+    private final byte id;
     private final String name;
 
-    public Selector(final int id, final String name) {
+    public Selector(final byte id, final String name) {
         this.id = id;
         this.name = name;
     }
 
-    public final int getId() {
+    public final byte getId() {
         return id;
     }
